Terminar el filtro avanzado

+++ Dominio/Producto.cs
@@ -15,6 +15,7 @@ namespace Dominio
         public Marca Marca { get; set; }
         public Categoria Categoria { get; set; }
         public string Imagen { get; set; }
-        public Double Precio { get; set; }
+        public decimal Precio { get; set; }
+
     }
 }
